pyproject.toml
[tool.poetry]
name = "cvcreator"
version = "1.0-alpha1"
description = "An automated tool for creating CVs on the fly."
authors = ["Jonathan Feinberg <jonathf@gmail.com>"]
license = "mit"
include = ["cvcreator/templates/*"]

[tool.poetry.dependencies]
python = ">=3.7"
PyYAML = "^5.4.1"
pydantic = "^1.8.1"
Jinja2 = "^2.11.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"

[tool.poetry.scripts]
_cvcreate = "cvcreator.__main__:main"
cvcreate = "cvcreator.cvcreate:main"
yaml2yaml = "cvcreator.yaml2yaml:main"
txt2yaml = "cvcreator.txt2yaml:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
